Changed pheromone path system
Added pheromone as independent class. Every object contains the next and the previous pheromones in the path.

ConsiderePheromone removed from Anttest.cs, converted to Sensepheromone where the overlapsphere is used.
Removed the arrays containing the "visible" pheromoneNodes.
Added Pheromone.cs with Pheromone class.
original Pheromone.cs now called PheromoneNode.cs. Contains the simplified node system.
Ant now does not have constant boxcollider to sense the pheromones. Instead, it uses an overlapsphere only when it needs to find a pheromone to follow. When it has one, it moves towards it. When it reaches it, it askes the pheromone for the next one and follows that one, etc.

Encountered an issue where the ant will accidentally stray from the pheromone path and be unable to return if it is above or below the ant. Suggested change: add AUX pheromone nodes that the ant releases when following the pheromone path. if the ant loses track of the pheromone path is can follow back the AUX nodes it left behind to find its way back.