Fixed error with getting adyacentcubes
GetGroup would return the wrong values, making it so cubeSurfaces' surfaceGroups would just be the cube value group. Fixed error both by making surfaceGroups values be true when above surface, and false when below (because only below would not include other surfaces in cube as also below in GetGroup) and fixing GetGroup output to be: the group. Believe it or not.