Implemented dig animation and fixed gravity
Added new state diggingAnim that signifies that the ant is busy digging. I also added a trigger to the animation, that calls the dig function. Could be good to expand on.

Right now once dig is called, the ant exits digstate before the animation finishes, but since passive waits a bit before doing anything, it works out. For now. Not sure 100% of this.

Also fixed gravity issue caused by the gravity not being locally turned on/of, so i moved the global gravity to worldGen and now the ant itself changes the rigidbody's usage of it.

Also added function Aling that makes the ant look at the digpoint before digging.