Food item, pickup and anim update
Added a corn mesh as a food item. Gave it a rigidbody and collider so as to have the ant see it. Updated ant animator board to include all movement animations while holding things, functioning through triggers to get in and out the holding state.

lowered ant drag considering previous removal of seeing other ants as surface a while back

added an emptyobject to the dolores prefab under the head bone. It is placed where the mouth is. This way, i can set an objects parent to the empty so that the ant is carrying it in its mouth and it moves with the head. It also has a specific rotation to make it look more natural. When picking up an object, i found the most simple thing to do be remove the rigidbody from the picked up object, and then give it a new one when it gets dropped.

ant.cs:
- removed the havegoal bool in favor of just having the next goal be negative inf when not assigned, as well as some functions to regulate it.
- Added the emptyobject as carriedObject to the script.
- added instance of the new class objective
- Added and implemented sensegeneral: senses food, then digpoints, then pheromones. the first it finds it sets the ai to that mode.
- created pickupaction: function that is called by the pick up anim. Since the digging uses the same anim (for now), the function acts differently depending on whether the ant is digging or picking up. Digging: calls the dig function of the DigPoint object in the objective. Picking up: removes rigidbody of food item in objective, then sets its parent to the emptyobject under ant head bone.
- PutDownAction: called by the put down anim. It makes the ant release all the childs of the emptyobject and give them rigidbodies

Camera: added the option to let go when controlling ant

new class objective: contains the pos of the objective, the mode (food, dig, pos) and the relevant item depending on its mode. Gave other classes ways to create it.