Edited FluedBehaviorTree
Edited the behaviour tree to get it to not relaunch every node on every tick. Just by having both selector and sequence nodes remember the nodes they are on and not start their first nodes again when tick is used till they get to a running node (or their fail/success nodes) and instead just have them get back to the running node again immediately, I gave it the functionality I needed.