Revamped Behaviour tree
Now there is a main sequence tree that starts with checking if there is not an objective to follow. If there isnt, it gets the ant a new one.

Then every next branch checks if the relevant task is the branch's. If so it enters branch and does the stuff. Branches must check if they are valid, and now when checking validity and getting false, the objective is removed so the first branch is entered on the next run.

Added senseTask, which does the senseFood and Sense Digpoint at the same time in order to use less resources.

Renamed objective class to Task. Added a queue of tasks that ants will check as priority tasks to choose. Usage not implemented yet.