Added new class: NEST
Nest represents all the parts that form the dug out nest. These parts are used to decide whether or not somewhere in the map is in the nest or not.

InNest uses the  marchingValue function used to determine the digPoints to check if the point is in the nest or not. There is a slight increase in distance permited, in order to be sure that it includes all surfaces that have been dug out (this way one point cant be too far). The problem is that the distance increase is proportional to the size (ex. 1.2 times) instead of a flat increase.