Fixed digPoints being left in the air
Added extra check, where digpoints will look if they are separated from main terrain with Separate() every 100 updates. It travels to adyacent digPointsDatas till it reaches a underground point without a digpointdata. If it doesnt find any, it returns true: its separate.

When the digpoint sees this, it digs itself.