Added ant and corn counters to UI
Made FlyCamera update the values whenever it notices the static bools updateantcount and updatecorncount activate.

Only counts cornCobs in foodChamber. If no foodchamber, it displays ?.
Only counts ants when there is an eggChamber. Otherwise it displays ?.
For now counts eggs as ants.

Updates corncount and antcount when ant grabs/places object.
Updates antCount when ant queen gives birth and when an ant is destroyed